feat: add config for auth module
