+++ src/app/(admin)/auth/_components/form-signup.tsx
@@ -1,0 +1,91 @@
+
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form"
+import Link from "next/link"
+import { User2Icon } from "lucide-react"
+import { Input } from "@/components/ui/input"
+import { Button } from "@/components/ui/button"
+import { useFormSignup } from "@auth/_hooks/form-signup"
+import { Avatar, AvatarFallback } from "@/components/ui/avatar"
+import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/components/ui/card'
+
+
+export default function FormChangePassword() {
+
+  const { form, onSubmit } = useFormSignup()
+
+  return (
+    <Card className="page-login shadow-none border-0">
+      <CardHeader className="flex flex-col justify-start items-center text-center">
+        <Avatar className="w-14 h-14 bg-transparent">
+          <AvatarFallback className="bg-transparent ring-1 ring-inset ring-primary/10">
+            <Avatar className="w-10 h-10 bg-transparent">
+              <AvatarFallback className="bg-transparent ring-1 ring-inset ring-primary/80">
+                <User2Icon />
+              </AvatarFallback>
+            </Avatar>
+          </AvatarFallback>
+        </Avatar>
+        <CardTitle>Create new account</CardTitle>
+        <CardDescription>Enter your email to create new account</CardDescription>
+      </CardHeader>
+      <CardContent className="pb-6">
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)} className="w-full flex flex-col justify-start items-stretch space-y-6">
+            <FormField
+              control={form.control}
+              name="email"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Email</FormLabel>
+                  <FormControl>
+                    <Input placeholder="shadcn@email.com" {...field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <FormField
+              control={form.control}
+              name="password"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Password</FormLabel>
+                  <FormControl>
+                    <Input placeholder="********" {...field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <FormField
+              control={form.control}
+              name="confirmPassword"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Confirm password</FormLabel>
+                  <FormControl>
+                    <Input placeholder="********" {...field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <div className="flex flex-col space-y-2">
+              <Button type="submit">Signup with email</Button>
+              <Button type="submit" variant={'secondary'}>Signup with Google</Button>
+              <Button type="submit" variant={'secondary'}>Signup with Github</Button>
+              <Button type="button" variant={'link'}><Link href={'/auth/login'}>Found an account? Back to login</Link></Button>
+            </div>
+          </form>
+        </Form>
+      </CardContent>
+    </Card>
+  )
+}
