+++ src/app/(admin)/auth/_hooks/form-reset-password.tsx
@@ -1,0 +1,35 @@
+import { z } from "zod"
+import { useForm } from "react-hook-form"
+import { toast } from "@/hooks/use-toast"
+import { zodResolver } from "@hookform/resolvers/zod"
+
+export function useFormResetPassword() {
+
+  const FormSchema = z.object({
+    email: z.string().email(),
+  })
+
+  const form = useForm<z.infer<typeof FormSchema>>({
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      email: "",
+    },
+  })
+
+  function onSubmit(data: z.infer<typeof FormSchema>) {
+    console.log(data)
+    toast({
+      title: "You submitted the following values:",
+      description: (
+        <pre className="mt-2 w-[340px] rounded-md bg-primary p-4">
+          <code className="text-white">{JSON.stringify(data, null, 2)}</code>
+        </pre>
+      )
+    })
+  }
+
+  return {
+    form,
+    onSubmit
+  }
+}
