+++ src/app/(admin)/auth/change-password/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import FormChangePassword from "@auth/_components/form-change-password"
+
+export default function PageChangePassword() {
+  return <FormChangePassword />
+}
