+++ src/app/(admin)/auth/login/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import FormLogin from "@auth/_components/form-login"
+
+export default function PageLogin() {
+  return <FormLogin />
+}
