+++ src/app/(admin)/auth/reset-password/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import FormResetPassword from "@auth/_components/form-reset-password"
+
+export default function PageResetPassword() {
+  return <FormResetPassword />
+}
