+++ src/app/(admin)/auth/signup/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import FormSignup from "@auth/_components/form-signup"
+
+export default function PageSignup() {
+  return <FormSignup />
+}
